requirements.txt
git+https://github.com/smarnach/pyexiftool.git
arrow
pillow
pexif
tzwhere
